internal: export NativeEndian
